feat: move selectable periods hook file
