+++ src/webapp/hooks/useSelectablePeriods.ts
